M src/server/j2se/com/sun/sgs/impl/service/data/store/net/DataStoreClient.java
- Cosmetic changes.

M src/server/j2se/com/sun/sgs/impl/service/data/store/net/DataStoreServer.java

- Modify createTransaction to throw IllegalArgumentException if the
  timeout is zero, not just negative.

M src/server/j2se/com/sun/sgs/impl/service/data/store/net/DataStoreServerImpl.java

- Correct the default max txn timeout.
- Cosmetic changes.
- Modify createTransaction to check for illegal timeout values.

M src/server/j2se/com/sun/sgs/impl/service/transaction/TransactionCoordinatorImpl.java

- Cosmetic changes.

M test/server/j2se/com/sun/sgs/test/impl/service/data/store/net/TestDataStoreClient.java

- Test zero timeout.

M test/server/j2se/com/sun/sgs/test/impl/service/data/store/net/TestDataStoreServerImpl.java

- Test bad transaction timeouts.

M test/shared/j2se/com/sun/sgs/test/impl/sharedutil/TestPropertiesWrapper.java

- Cosmetic changes.

+++ src/server/j2se/com/sun/sgs/impl/service/data/store/net/DataStoreClient.java
@@ -35,7 +35,6 @@ import java.util.logging.Logger;
  *	</b></code><br>
  *	<i>Default:</i> {@code 100}
  *
- *
  * <dd style="padding-top: .5em">The number of object IDs to allocate at one
  *	time.  This value must be greater than {@code 0}. <p>
  *
@@ -235,8 +234,8 @@ public final class DataStoreClient
 	    ALLOCATION_BLOCK_SIZE_PROPERTY, DEFAULT_ALLOCATION_BLOCK_SIZE,
 	    1, Integer.MAX_VALUE);
 	maxTxnTimeout = wrappedProps.getLongProperty(
-	    MAX_TXN_TIMEOUT_PROPERTY, DEFAULT_MAX_TXN_TIMEOUT,
-	    (long) 0, Long.MAX_VALUE);
+	    MAX_TXN_TIMEOUT_PROPERTY, DEFAULT_MAX_TXN_TIMEOUT, 0,
+	    Long.MAX_VALUE);
 	server = getServer();
     }
